pyproject.toml
[project]
name = "propuestas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "weasyprint>=60.0",
    "jinja2>=3.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "openai>=1.0.0",
    "boto3>=1.34.0",
    "python-multipart>=0.0.6",
    "rollbar>=0.16.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "google-auth>=2.40.3",
]
